requirements.txt
SQLAlchemy==1.0.12
psycopg2==2.6.2
awesome-slugify
